+++ web/spectron/react-css/index.ts
@@ -1,0 +1,12 @@
+import {SpectronMain2} from '../../js/test/SpectronMain2';
+
+SpectronMain2.create().run(async state => {
+
+    state.window.loadURL(`file://${__dirname}/content.html`);
+
+    // await state.testResultWriter.write(true);
+
+});
+
+
+